% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_curl_proxy.R
\name{configure_curl_proxy}
\alias{configure_curl_proxy}
\title{Set internet proxy when on DHSC estate}
\usage{
configure_curl_proxy(target_url = "https://www.google.co.uk")
}
\arguments{
\item{target_url}{The url to get the proxy settings for, defaults to \url{https://www.google.co.uk}.}
}
\description{
Configure proxy settings for functions based on curl using Windows settings.
Needed when on the DHSC estate using a wired connection due to corporate
firewall.
}
\details{
While R uses the \code{wininet} download method that correctly obtains
proxy details from Windows via the WinINet functions, other packages
such as \pkg{httr} are based on \pkg{curl} and so need the proxy settings
configured when using a wired connection on the DHSC estate due to
the corporate firewall. This function uses \code{curl::ie_get_proxy_for_url}
to read these settings and then applies them using \code{httr::set_config}.
}
\examples{
configure_curl_proxy()
httr::content(httr::GET("https://api.ipify.org"), encoding = "UTF-8")
}
