% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_git_proxy.R
\name{configure_git_proxy}
\alias{configure_git_proxy}
\title{Set git http and https proxies for DHSC}
\usage{
configure_git_proxy()
}
\description{
Configure http and https proxy settings for our corporate firewall when using
git on the DHSC estate through a wired connection.
}
\details{
Function uses \code{curl::ie_get_proxy_for_url} to read the Windows proxy
settings for \url{http://github.com} and \url{https://github.com} then uses
these to configure git with the \code{usethis::use_git_config} helper
function.

NOTE: This requires that projects are configured / cloned from
GitHub using the HTTPS (not SSH) link.

The function also needs to disable SSL verification but this
represents a minimal security risk when using the corporate firewall.
}
